files database operations moved into own file.
